add createEventListener utility
