+++ packages/nimble-components/src/menu-button/tests/menu-button.spec.ts
@@ -15,6 +15,7 @@ import {
     processUpdates,
     waitForUpdatesAsync
 } from '../../testing/async-helpers';
+import { createEventListener } from '../../utilities/tests/component';
 
 class TestSlottedElement extends FoundationElement {}
 const composedTestSlottedElement = TestSlottedElement.compose({
@@ -32,28 +33,6 @@ async function setup(): Promise<Fixture<MenuButton>> {
 
 async function slottedSetup(): Promise<Fixture<TestSlottedElement>> {
     return fixture(composedTestSlottedElement());
-}
-
-/** A helper function to abstract adding a `toggle` event listener, spying
- * on the event being called, and removing the event listener. The returned promise
- * should be resolved prior to completing a test.
- */
-function createToggleListener(menuButton: MenuButton): {
-    promise: Promise<void>,
-    spy: jasmine.Spy
-} {
-    const spy = jasmine.createSpy();
-    return {
-        promise: new Promise(resolve => {
-            const handler = (...args: unknown[]): void => {
-                menuButton.removeEventListener('toggle', handler);
-                spy(...args);
-                resolve();
-            };
-            menuButton.addEventListener('toggle', handler);
-        }),
-        spy
-    };
 }
 
 /** A helper function to abstract adding a `beforetoggle` event listener, spying
@@ -257,7 +236,7 @@ describe('MenuButton', () => {
 
         it("should fire 'toggle' event when the menu is opened", async () => {
             await connect();
-            const toggleListener = createToggleListener(element);
+            const toggleListener = createEventListener(element, 'toggle');
             element.open = true;
             await toggleListener.promise;
             expect(toggleListener.spy).toHaveBeenCalledTimes(1);
@@ -273,7 +252,7 @@ describe('MenuButton', () => {
         it("should fire 'toggle' event when the menu is closed", async () => {
             element.open = true;
             await connect();
-            const toggleListener = createToggleListener(element);
+            const toggleListener = createEventListener(element, 'toggle');
             element.open = false;
             await toggleListener.promise;
             expect(toggleListener.spy).toHaveBeenCalledTimes(1);
@@ -288,7 +267,7 @@ describe('MenuButton', () => {
 
         it("should fire 'beforetoggle' event before the menu opens", async () => {
             await connect();
-            const toggleListener = createToggleListener(element);
+            const toggleListener = createEventListener(element, 'toggle');
             const beforeToggleListener = createBeforeToggleListener(
                 element,
                 false,
@@ -315,7 +294,7 @@ describe('MenuButton', () => {
         it("should fire 'beforetoggle' event before the menu is closed", async () => {
             element.open = true;
             await connect();
-            const toggleListener = createToggleListener(element);
+            const toggleListener = createEventListener(element, 'toggle');
             const beforeToggleListener = createBeforeToggleListener(
                 element,
                 true,
@@ -370,7 +349,7 @@ describe('MenuButton', () => {
                     return;
                 }
 
-                const toggleListener = createToggleListener(menuButton);
+                const toggleListener = createEventListener(menuButton, 'toggle');
                 menuButton.open = true;
                 await toggleListener.promise;
             }
@@ -387,7 +366,7 @@ describe('MenuButton', () => {
             it('should open the menu and focus first menu item when the toggle button is clicked', async () => {
                 await connect();
                 const menuButton = configuration.getMenuButton(element);
-                const toggleListener = createToggleListener(menuButton);
+                const toggleListener = createEventListener(menuButton, 'toggle');
                 menuButton.toggleButton!.control.click();
                 expect(menuButton.open).toBeTrue();
                 await toggleListener.promise;
@@ -397,7 +376,7 @@ describe('MenuButton', () => {
             it("should open the menu and focus first menu item when 'Enter' is pressed while the toggle button is focused", async () => {
                 await connect();
                 const menuButton = configuration.getMenuButton(element);
-                const toggleListener = createToggleListener(menuButton);
+                const toggleListener = createEventListener(menuButton, 'toggle');
                 const event = new KeyboardEvent('keypress', {
                     key: keyEnter
                 } as KeyboardEventInit);
@@ -410,7 +389,7 @@ describe('MenuButton', () => {
             it("should open the menu and focus first menu item when 'Space' is pressed while the toggle button is focused", async () => {
                 await connect();
                 const menuButton = configuration.getMenuButton(element);
-                const toggleListener = createToggleListener(menuButton);
+                const toggleListener = createEventListener(menuButton, 'toggle');
                 const event = new KeyboardEvent('keypress', {
                     key: keySpace
                 } as KeyboardEventInit);
@@ -423,7 +402,7 @@ describe('MenuButton', () => {
             it('should open the menu and focus first menu item when the down arrow is pressed while the toggle button is focused', async () => {
                 await connect();
                 const menuButton = configuration.getMenuButton(element);
-                const toggleListener = createToggleListener(menuButton);
+                const toggleListener = createEventListener(menuButton, 'toggle');
                 const event = new KeyboardEvent('keydown', {
                     key: keyArrowDown
                 } as KeyboardEventInit);
@@ -436,7 +415,7 @@ describe('MenuButton', () => {
             it('should open the menu and focus last menu item when the up arrow is pressed while the toggle button is focused', async () => {
                 await connect();
                 const menuButton = configuration.getMenuButton(element);
-                const toggleListener = createToggleListener(menuButton);
+                const toggleListener = createEventListener(menuButton, 'toggle');
                 const event = new KeyboardEvent('keydown', {
                     key: keyArrowUp
                 } as KeyboardEventInit);
@@ -567,7 +546,7 @@ describe('MenuButton', () => {
                     return;
                 }
 
-                const toggleListener = createToggleListener(menuButton);
+                const toggleListener = createEventListener(menuButton, 'toggle');
                 menuButton.open = true;
                 await toggleListener.promise;
             }
@@ -584,7 +563,7 @@ describe('MenuButton', () => {
             it('should transition to the open state when the toggle button is clicked', async () => {
                 await connect();
                 const menuButton = configuration.getMenuButton(element);
-                const toggleListener = createToggleListener(menuButton);
+                const toggleListener = createEventListener(menuButton, 'toggle');
                 menuButton.toggleButton!.control.click();
                 expect(menuButton.open).toBeTrue();
                 await toggleListener.promise;
@@ -601,7 +580,7 @@ describe('MenuButton', () => {
             it("should transition to the open state when 'Enter' is pressed while the toggle button is focused", async () => {
                 await connect();
                 const menuButton = configuration.getMenuButton(element);
-                const toggleListener = createToggleListener(menuButton);
+                const toggleListener = createEventListener(menuButton, 'toggle');
                 const event = new KeyboardEvent('keypress', {
                     key: keyEnter
                 } as KeyboardEventInit);
@@ -621,7 +600,7 @@ describe('MenuButton', () => {
             it("should transition to the open state when 'Space' is pressed while the toggle button is focused", async () => {
                 await connect();
                 const menuButton = configuration.getMenuButton(element);
-                const toggleListener = createToggleListener(menuButton);
+                const toggleListener = createEventListener(menuButton, 'toggle');
                 const event = new KeyboardEvent('keypress', {
                     key: keySpace
                 } as KeyboardEventInit);
@@ -641,7 +620,7 @@ describe('MenuButton', () => {
             it('should transition to the open state when the down arrow is pressed while the toggle button is focused', async () => {
                 await connect();
                 const menuButton = configuration.getMenuButton(element);
-                const toggleListener = createToggleListener(menuButton);
+                const toggleListener = createEventListener(menuButton, 'toggle');
                 const event = new KeyboardEvent('keydown', {
                     key: keyArrowDown
                 } as KeyboardEventInit);
@@ -661,7 +640,7 @@ describe('MenuButton', () => {
             it('should transition to the open state when the up arrow is pressed while the toggle button is focused', async () => {
                 await connect();
                 const menuButton = configuration.getMenuButton(element);
-                const toggleListener = createToggleListener(menuButton);
+                const toggleListener = createEventListener(menuButton, 'toggle');
                 const event = new KeyboardEvent('keydown', {
                     key: keyArrowUp
                 } as KeyboardEventInit);
